// src/app/service/api/get/ticket-detail.service.ts
import { Injectable } from '@angular/core';
import { HttpClient, HttpHeaders, HttpParams } from '@angular/common/http';

import { environment } from '../../../../environments/environment';

@Injectable({
  providedIn: 'root'
})
export class TicketDetailService {
  private url: string = environment.managerHost + 'export_opd_data/';
  constructor(
    private httpClient: HttpClient
  ) { }

  getTicketDetail(token, ticketId){
    let headers = new HttpHeaders({
      'Content-Type': 'application/json',
      'Authorization': "Bearer "+token,
    });

    return this.httpClient.get<Response>(this.url+ticketId, { headers: headers })
  }
}
export interface Response{
  code: number;
  message: string;
  data: Ticket;
}

export interface Ticket {
  ID: number;
  hospital_user_id: number;
  hospital_employee_name: string;
  hospital_phone_number: string;
  patient_phone_numb: string;
  insmart_user_id: number;
  ill_cause: number;
  isurance_id: number;
  card_no: string;
  cmnd: string;
  fullname: string;
  dob: string;
  policy_no: string;
  diag_note: string;
  maximum_claim_value: number;
  is_apply_social_insurance: number;
  social_insurance_id: string;
  files: string;
  costs: string;
  note: string;
  insmart_status: string;
  hospital_status: string;
  created_at: string;
  updated_at: string;
  deleted_at: string;
  hospital_updated_at: string;
  detailTicket: Comments;
  insmartCosts: any;
  hospitalCosts: any;
  reasonReject: string;
  lastAddition: string;
  insmartUpdatedCostAt: string;
  insmartRejectAt: string;
  hospitalUpdatedCostAt: string;
}
export interface Comments extends Array<Comment>{};
interface Comment{
  ID: number;
  content: any;
  created_at: string;
  hospital_user_id: number;
  insmart_user_id: number;
  ticket_id: number;
  type: string;
}